+++ Source/Krypton Components/Krypton.Utilities/Global/Globals.cs
@@ -15,4 +15,5 @@ global using System.Drawing;
 global using System.Windows.Forms;
 global using System.Windows.Forms.Design;
 
+global using Krypton.Ribbon;
 global using Krypton.Toolkit;